LOY-1231: code base - mock Database, declare TEST profile

+++ internal/fileawardpoint/file_award_point.go
@@ -5,13 +5,13 @@ import "time"
 // FileAwardPoint is model of table `file_award_point`
 type FileAwardPoint struct {
 	Id                int       `json:"id"`
-	MerchantId        int       `json:"merchant_id"`
+	MerchantId        int64     `json:"merchant_id"`
 	DisplayName       string    `json:"display_name"`
 	FileUrl           string    `json:"file_url"`
 	ResultFileUrl     string    `json:"result_file_url"`
-	Status            int       `json:"status"`
-	StatsTotalRow     int       `json:"stats_total_row"`
-	StatsTotalSuccess int       `json:"stats_total_success"`
+	Status            int16     `json:"status"`
+	StatsTotalRow     int16     `json:"stats_total_row"`
+	StatsTotalSuccess int16     `json:"stats_total_success"`
 	CreatedAt         time.Time `json:"created_at"`
 	UpdatedAt         time.Time `json:"updated_at"`
 	CreatedBy         string    `json:"created_by"`
